+++ apps/web-client/src/lib/db/cr-sqlite/__tests__/reconciliation-order.test.ts
@@ -1,6 +1,6 @@
-import { describe, it, expect, beforeEach } from "vitest";
+import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
 
-import { type DB } from "../types";
+import { type DB, type ProcessedOrderLine } from "../types";
 
 import { getRandomDb, createCustomerOrders } from "./lib";
 
@@ -10,10 +10,14 @@ import {
 	getReconciliationOrder,
 	addOrderLinesToReconciliationOrder,
 	finalizeReconciliationOrder,
-	getReconciliationOrderLines
+	getReconciliationOrderLines,
+	processOrderDelivery,
+	sortLinesBySupplier
 } from "../order-reconciliation";
 import { createSupplierOrder, getPlacedSupplierOrders, getPossibleSupplierOrderLines } from "../suppliers";
 import { getCustomerBooks } from "../customers";
+
+import {} from "../order-reconciliation";
 
 // TODO: this needs some work... leaving till reconcilation wiring in effort/updates
 describe("Reconciliation order creation", () => {
@@ -251,5 +255,181 @@ describe("Reconciliation order creation", () => {
 				`Reconciliation order ${reconOrderId} is already finalized`
 			);
 		});
+	});
+});
+
+describe("Misc helpers", () => {
+	afterEach(() => {
+		vi.clearAllMocks();
+	});
+	it("should process when delivery matches order exactly", () => {
+		const scannedBooks = [{ isbn: "123", title: "Book 1", authors: "Author 1", price: 10, quantity: 2 }];
+
+		const placedOrderLines = [
+			{
+				isbn: "123",
+				title: "Book 1",
+				authors: "Author 1",
+				price: 10,
+				quantity: 2,
+				supplier_name: "Supplier 1",
+				id: 1,
+				supplier_id: 1,
+				total_book_number: 1,
+				supplier_order_id: 1,
+				total_price: 10,
+				created: Date.now()
+			}
+		];
+
+		const result = processOrderDelivery(scannedBooks, placedOrderLines);
+		expect(result).toEqual({
+			processedLines: [
+				{
+					isbn: "123",
+					authors: "Author 1",
+					deliveredQuantity: 2,
+					orderedQuantity: 2,
+					price: 10,
+					quantity: 2,
+					title: "Book 1"
+				}
+			],
+			unmatchedBooks: []
+		});
+	});
+
+	it("should handle partial delivery", () => {
+		const scannedBooks = [{ isbn: "123", title: "Book 1", authors: "Author 1", price: 10, quantity: 1 }];
+
+		const placedOrderLines = [
+			{
+				isbn: "123",
+				title: "Book 1",
+				authors: "Author 1",
+				price: 10,
+				quantity: 2,
+				supplier_name: "Supplier 1",
+				id: 1,
+				supplier_id: 1,
+				total_book_number: 1,
+				supplier_order_id: 1,
+				total_price: 10,
+				created: Date.now()
+			}
+		];
+
+		const result = processOrderDelivery(scannedBooks, placedOrderLines);
+		expect(result).toEqual({
+			processedLines: [
+				{
+					isbn: "123",
+					authors: "Author 1",
+					deliveredQuantity: 1,
+					orderedQuantity: 2,
+					price: 10,
+					quantity: 1,
+					title: "Book 1"
+				}
+			],
+			unmatchedBooks: []
+		});
+	});
+
+	it("should handle over-delivery", () => {
+		const scannedBooks = [{ isbn: "123", title: "Book 1", authors: "Author 1", price: 10, quantity: 3 }];
+
+		const placedOrderLines = [
+			{
+				isbn: "123",
+				title: "Book 1",
+				authors: "Author 1",
+				price: 10,
+				quantity: 2,
+				supplier_name: "Supplier 1",
+				id: 1,
+				supplier_id: 1,
+				total_book_number: 1,
+				supplier_order_id: 1,
+				total_price: 10,
+				created: Date.now()
+			}
+		];
+
+		const result = processOrderDelivery(scannedBooks, placedOrderLines);
+		expect(result).toEqual({
+			processedLines: [
+				{
+					isbn: "123",
+					authors: "Author 1",
+					deliveredQuantity: 3,
+					orderedQuantity: 2,
+					price: 10,
+					quantity: 3,
+					title: "Book 1"
+				}
+			],
+			unmatchedBooks: []
+		});
+	});
+
+	it("should handle unordered books", () => {
+		const scannedBooks = [{ isbn: "456", title: "Book 2", authors: "Author 2", price: 15, quantity: 2 }];
+
+		const placedOrderLines = [
+			{
+				isbn: "123",
+				title: "Book 1",
+				authors: "Author 1",
+				price: 10,
+				quantity: 2,
+				supplier_name: "Supplier 1",
+				id: 1,
+				supplier_id: 1,
+				total_book_number: 1,
+				supplier_order_id: 1,
+				total_price: 10,
+				created: Date.now()
+			}
+		];
+
+		const result = processOrderDelivery(scannedBooks, placedOrderLines);
+		expect(result).toEqual({
+			processedLines: [],
+			unmatchedBooks: [
+				expect.objectContaining({
+					authors: "Author 1",
+					id: 1,
+					isbn: "123",
+					price: 10,
+					quantity: 2,
+					supplier_id: 1,
+					supplier_name: "Supplier 1",
+					supplier_order_id: 1,
+					title: "Book 1",
+					total_book_number: 1,
+					total_price: 10
+				}),
+				expect.objectContaining({
+					authors: "Author 2",
+					isbn: "456",
+					price: 15,
+					quantity: 2,
+					title: "Book 2"
+				})
+			]
+		});
+	});
+	it("should group order lines by supplier", () => {
+		const orderLines = [
+			{ supplier_name: "Supplier 1", isbn: "123" },
+			{ supplier_name: "Supplier 2", isbn: "456" },
+			{ supplier_name: "Supplier 1", isbn: "789" }
+		] as ProcessedOrderLine[];
+
+		const result = sortLinesBySupplier(orderLines);
+		expect(Object.keys(result)).toHaveLength(2);
+		expect(result["Supplier 1"]).toHaveLength(2);
+		expect(result["Supplier 2"]).toHaveLength(1);
 	});
 });
