+++ apps/web-client/src/routes/orders/suppliers/order/[id]/+page.ts
@@ -6,6 +6,7 @@ export const load: PageLoad = async ({ parent, params }) => {
 
 	const placedOrder = await getPlacedSupplierOrderLines(ordersDb, parseInt(params.id));
 
+	console.log({ placedOrder });
 	return { placedOrder };
 };
